# docs/openapi.yaml
openapi: 3.0.3
info:
  title: oServ API (Next.js)
  version: "1.0.0"
  description: |
    Documentação da API do sistema **oServ** após a migração para **Next.js Fullstack**.  
    Todas as rotas de API estão sob o prefixo `/api` e seguem a arquitetura multiempresa descrita nos documentos do projeto.  
    As principais rotas listadas no arquivo `README.md` incluem autenticação, gerenciamento da empresa, perfis, cadastros, ordens de serviço, comentários e notificações【890973672335846†L291-L352】.  
    Esta especificação também considera as entidades e campos sugeridos na proposta de arquitetura SaaS multiempresa【169360353382499†L31-L120】.
servers:
  - url: http://localhost:3000
    description: Servidor local de desenvolvimento
paths:
...
  /api/service-orders:
    get:
      summary: Listar ordens de serviço
      description: |
        Retorna todas as ordens de serviço da empresa autenticada.  
        Filtros opcionais permitem listar por status ou por responsável【890973672335846†L331-L343】.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [OPEN, IN_PROGRESS, COMPLETED, CANCELLED]
          description: Filtrar pelo status da ordem.
        - name: responsibleEmployeeId
          in: query
          required: false
          schema:
            type: string
          description: Filtrar ordens pelo ID do funcionário responsável.
      responses:
        "200":
          description: Lista de ordens de serviço
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceOrder'
        "401":
          description: Não autenticado
    post:
      summary: Cadastrar ordem de serviço
      description: |
        Cria uma nova ordem de serviço.  
        Endpoint listado em `README.md`【890973672335846†L331-L337】.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceOrderCreateRequest'
      responses:
        "201":
          description: Ordem de serviço criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrder'
        "400":
          description: Dados inválidos
        "401":
          description: Não autenticado
    patch:
      summary: Atualizar ordem de serviço
      description: |
        Atualiza dados de uma ordem de serviço existente, como status ou informações adicionais.  
        Endpoint listado em `README.md`【890973672335846†L331-L337】.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceOrderUpdateRequest'
      responses:
        "200":
          description: Ordem de serviço atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceOrder'
        "400":
          description: Dados inválidos
        "401":
          description: Não autenticado
    delete:
      summary: Excluir ordem de serviço
      description: |
        Remove uma ordem de serviço pelo ID informado na query string.  
        Endpoint listado em `README.md`【890973672335846†L331-L337】.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: Identificador da ordem de serviço.
      responses:
        "204":
          description: Ordem removida
        "401":
          description: Não autenticado
...
components:
  schemas:
...
    OrderStatus:
      type: string
      description: Status de uma ordem de serviço.
      enum:
        - OPEN
        - IN_PROGRESS
        - COMPLETED
        - CANCELLED
...
    ServiceOrder:
      type: object
      description: Entidade ordem de serviço que organiza solicitações de clientes.
      properties:
        id:
          type: string
        companyId:
          type: string
        customerId:
          type: string
        serviceIds:
          type: array
          items:
            type: string
          description: Lista de IDs de serviços incluídos na ordem.
        responsibleEmployeeId:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        title:
          type: string
        description:
          type: string
        total:
          type: number
          format: double
        openedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - customerId
        - serviceIds
        - status
        - openedAt
...
    ServiceOrderCreateRequest:
      type: object
      description: Dados necessários para criar uma ordem de serviço.
      properties:
        customerId:
          type: string
        serviceIds:
          type: array
          items:
            type: string
        title:
          type: string
        description:
          type: string
        total:
          type: number
          format: double
        dueDate:
          type: string
          format: date-time
        responsibleEmployeeId:
          type: string
      required:
        - customerId
        - serviceIds
        - title
    ServiceOrderUpdateRequest:
      type: object
      description: Campos permitidos para atualizar uma ordem de serviço.
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        title:
          type: string
        description:
          type: string
        total:
          type: number
          format: double
        dueDate:
          type: string
          format: date-time
        responsibleEmployeeId:
          type: string
...